+++ ref-exchange/release_notes.md
@@ -1,8 +1,12 @@
 # Release Notes
 
-### Version 1.4.2
+### Version 1.4.3
 1. Let both guardians and owner can remove exchange liquidity to owner inner account by remove_exchange_fee_liquidity;
 2. Let both guardians and owner can withdraw owner token to owner wallet by withdraw_owner_token;
+
+### Version 1.4.2
+1. Let owner can retrieve unmanaged NEP-141 tokens in contract account;
+2. support withdraw token's full amount in inner-account with 0 in amount parameter;
 
 ### Version 1.4.1
 1. Introduce Stable-Swap-Pool;